Rename main class (#2)
* Add DEBUG logging to the API

This commit adds debug logs for the request and response. This can be
useful later for debugging purposes when this SDK is being used.

* Rename main class

This commit renames the main class that we use from `Api` to `Nightfall`
because it makes it a bit more clear about what is going on when you are
using this library in another module.