Implement dict instead of string
The scan() and make_payload() functions now accepts lists of dicts
instead of lists of strings.

In make_payload() If the total number of items is less than MAX_NUM_ITEMS
we return a list of dicts. If the sum of items is larger than MAX_PAYLOAD_SIZE
we return multiple lists of dicts. If the sum of an individual item
is larger than MAX_PAYLOAD_SIZE we return multiple lists of dicts
where the dict value is split with the id being retained for
later processing.

This is beneficial for implementation because it makes it easier to keep
track of where a specific string came from. This is especially true in
cases where we had to split a string due to API size limitations.

Lastly, we refactored the test suite and split up unit and integration
tests.

Unit tests found at test_api.py now use mock requests and we created a
new module called test_integrations.py which contains real requests to
the API.