Implement dict instead of string (#5)
The scan() and make_payload() functions now accepts lists of dicts
instead of lists of strings.

In make_payload() If the total number of items is less than MAX_NUM_ITEMS
we return a list of dicts. If the sum of items is larger than MAX_PAYLOAD_SIZE
we return multiple lists of dicts. If the sum of an individual item
is larger than MAX_PAYLOAD_SIZE we raise an exception.

This is beneficial for implementation because it makes it easier to keep
track of where a specific string came from.

Lastly, we refactored the test suite and split up unit and integration
tests.

Unit tests found at test_api.py now use mock requests and we created a
new module called test_integrations.py which contains real requests to
the API.